+++ cmp/grid/GridModel.js
@@ -20,6 +20,7 @@ import {
     cloneDeep,
     compact,
     defaults,
+    defaultsDeep,
     find,
     findLast,
     isArray,
@@ -115,6 +116,7 @@ export class GridModel {
     /**
      * @param {Object} c - GridModel configuration.
      * @param {Object[]} c.columns - {@link Column} or {@link ColumnGroup} configs
+     * @param {Object} [c.colDefaults] - Column configs to be set on all columns.  Merges deeply.
      * @param {(Store|Object)} [c.store] - a Store instance, or a config with which to create a
      *      Store. If not supplied, store fields will be inferred from columns config.
      * @param {boolean} [c.treeMode] - true if grid is a tree grid (default false).
@@ -152,6 +154,7 @@ export class GridModel {
     constructor({
         store,
         columns,
+        colDefaults = {},
         treeMode = false,
         showSummary = false,
         selModel,
@@ -197,6 +200,7 @@ export class GridModel {
 
         Object.assign(this, rest);
 
+        this.colDefaults = colDefaults;
         this.setColumns(columns);
         this.store = this.parseStore(store);
 
@@ -399,7 +403,7 @@ export class GridModel {
             colConfigs.some(c => !isPlainObject(c)),
             'GridModel only accepts plain objects for Column or ColumnGroup configs'
         );
-
+        
         const columns = colConfigs.map(c => this.buildColumn(c));
 
         this.validateColumns(columns);
@@ -557,7 +561,7 @@ export class GridModel {
     }
 
     buildColumn(c) {
-        return c.children ? new ColumnGroup(c, this) : new Column(c, this);
+        return c.children ? new ColumnGroup(c, this) : new Column(defaultsDeep({}, c, this.colDefaults), this);
     }
 
     //-----------------------